# unburdy_server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_ae-saas-basic_ae-saas-basic_internal_modules.ModuleInfo:
    properties:
      author:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      version:
        type: string
    type: object
...
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  models.ClientResponse:
    properties:
      admission_date:
        type: string
      alternative_email:
        type: string
      alternative_first_name:
        type: string
      alternative_last_name:
        type: string
      alternative_phone:
        type: string
      city:
        type: string
      contact_email:
        type: string
      contact_first_name:
        type: string
      contact_last_name:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      invoiced_individually:
        type: boolean
      last_name:
        type: string
      notes:
        type: string
      phone:
        type: string
      primary_language:
        type: string
      provider_approval_code:
        type: string
      referral_source:
        type: string
      status:
        type: string
      street_address:
        type: string
      therapy_title:
        type: string
      unit_price:
        type: number
      updated_at:
        type: string
      zip:
        type: string
    type: object
...
  models.CreateClientRequest:
    properties:
      admission_date:
        example: "2025-01-01"
        type: string
      alternative_email:
        example: johnny.d@example.com
        type: string
      alternative_first_name:
        example: Johnny
        type: string
      alternative_last_name:
        example: D
        type: string
      alternative_phone:
        example: "+0987654321"
        type: string
      city:
        example: New York
        type: string
      contact_email:
        example: jane.smith@example.com
        type: string
      contact_first_name:
        example: Jane
        type: string
      contact_last_name:
        example: Smith
        type: string
      contact_phone:
        example: "+1234567890"
        type: string
      date_of_birth:
        example: "1990-01-15"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      gender:
        example: male
        type: string
      invoiced_individually:
        example: false
        type: boolean
      last_name:
        example: Doe
        type: string
      notes:
        example: Additional notes about the client
        type: string
      phone:
        example: "+1234567890"
        type: string
      primary_language:
        example: English
        type: string
      provider_approval_code:
        example: PROV123456
        type: string
      referral_source:
        example: Doctor Smith
        type: string
      status:
        example: waiting
        type: string
      street_address:
        example: 123 Main Street
        type: string
      therapy_title:
        example: Cognitive Behavioral Therapy
        type: string
      unit_price:
        example: 150
        type: number
      zip:
        example: "12345"
        type: string
    required:
    - first_name
    - last_name
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
    type: object
...
  models.UpdateClientRequest:
    properties:
      admission_date:
        example: "2025-01-01"
        type: string
      alternative_email:
        example: johnny.d@example.com
        type: string
      alternative_first_name:
        example: Johnny
        type: string
      alternative_last_name:
        example: D
        type: string
      alternative_phone:
        example: "+0987654321"
        type: string
      city:
        example: New York
        type: string
      contact_email:
        example: jane.smith@example.com
        type: string
      contact_first_name:
        example: Jane
        type: string
      contact_last_name:
        example: Smith
        type: string
      contact_phone:
        example: "+1234567890"
        type: string
      date_of_birth:
        example: "1990-01-15"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      gender:
        example: male
        type: string
      invoiced_individually:
        example: false
        type: boolean
      last_name:
        example: Doe
        type: string
      notes:
        example: Additional notes about the client
        type: string
      phone:
        example: "+1234567890"
        type: string
      primary_language:
        example: English
        type: string
      provider_approval_code:
        example: PROV123456
        type: string
      referral_source:
        example: Doctor Smith
        type: string
      status:
        example: active
        type: string
      street_address:
        example: 123 Main Street
        type: string
      therapy_title:
        example: Cognitive Behavioral Therapy
        type: string
      unit_price:
        example: 150
        type: number
      zip:
        example: "12345"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@unburdy.com
    name: API Support
    url: https://unburdy.com/support
  description: Extended SaaS backend API built on AE SaaS Basic, adding client management
    functionality
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://unburdy.com/terms
  title: Unburdy Extended API
  version: "1.0"
paths:
  /admin/modules:
    get:
      consumes:
      - application/json
      description: Get information about all registered modules in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ae-saas-basic_ae-saas-basic_internal_modules.ModuleInfo'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get registered modules
      tags:
      - modules
  /admin/modules/{name}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific module by name
      parameters:
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ae-saas-basic_ae-saas-basic_internal_modules.ModuleInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get module details
      tags:
      - modules
...
  /clients:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all clients
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array of ClientResponse, total: int, page: int, limit:
            int'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: string'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Create a new client with first name, last name, and date of birth
      parameters:
      - description: Client data
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ClientResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new client
      tags:
      - clients
  /clients/{id}:
    delete:
      description: Soft delete a client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a client
      tags:
      - clients
    get:
      description: Retrieve a specific client by their ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClientResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a client by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Update an existing client's information
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client update data
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClientResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a client
      tags:
      - clients
  /clients/search:
    get:
      description: Search clients by first name or last name
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search clients
      tags:
      - clients
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...